+++ src/components/shared/Game/GameSummary/GameSummaryModal/GameSummaryHistoryPlayLine.vue
@@ -35,6 +35,16 @@
                 <RoleText class="text-center small cursor-text" :role="target.player.role.current"/>
             </div>
         </td>
+        <div v-if="actionConsequences.length" class="w-100 font-italic p-1">
+            <hr class="bg-dark my-1"/>
+            <div class="font-weight-bold ml-2">
+                <i class="fa fa-dice text-info mr-2"/>
+                <span v-html="`${$t('GameSummaryHistoryPlayLine.remarksOnPlay')} :`"/>
+            </div>
+            <ul class="mb-2">
+                <li v-for="actionConsequence of actionConsequences" :key="actionConsequence" v-html="actionConsequence"/>
+            </ul>
+        </div>
     </tr>
 </template>
 
@@ -64,6 +74,8 @@ import stutteringJudgeSVG from "@/assets/svg/roles/stuttering-judge.svg";
 import whiteWerewolfSVG from "@/assets/svg/roles/white-werewolf.svg";
 import RoleImage from "@/components/shared/Game/Role/RoleImage";
 import RoleText from "@/components/shared/Game/Role/RoleText";
+import { insertIf } from "@/helpers/functions/Array";
+import { areAllAdditionalCardsWerewolves } from "@/helpers/functions/Game";
 
 export default {
     name: "GameSummaryHistoryPlayLine",
@@ -133,20 +145,38 @@ export default {
                     return this.$t(`GameSummaryHistoryPlayLine.actions.witch.use-potion.life`);
                 }
                 return this.$t(`GameSummaryHistoryPlayLine.actions.witch.use-potion.death`);
-            } else if (action === "eat" && source === "werewolves" && targets[0].isInfected) {
+            } else if (action === "eat" && source.name === "werewolves" && targets[0].isInfected) {
                 return this.$t(`GameSummaryHistoryPlayLine.actions.vile-father-of-wolves.infect`);
             }
             return this.$t(`GameSummaryHistoryPlayLine.actions.${source.name}.${action}`);
         },
         actionSource() {
             const { action, source, targets } = this.gameHistoryEntry.play;
-            if (action === "eat" && source === "werewolves" && targets[0].isInfected) {
+            if (action === "eat" && source.name === "werewolves" && targets[0].isInfected) {
                 return {
                     name: "vile-father-of-wolves",
                     players: [this.game.vileFatherOfWolvesPlayer],
                 };
             }
             return source;
+        },
+        actionConsequences() {
+            const { options, thiefAdditionalCards } = this.game;
+            const { action, source, targets, votesResult, doesJudgeRequestAnotherVote } = this.gameHistoryEntry.play;
+            const consequencesText = "GameSummaryHistoryPlayLine.consequences";
+            return [
+                ...insertIf(action === "protect" && targets[0].player.currentRole === "little-girl" && !options.roles.littleGirl.isProtectedByGuard,
+                    this.$t(`${consequencesText}.littleGirlNotProtected`)),
+                ...insertIf(action === "vote" && votesResult === "need-settlement", this.$t(`${consequencesText}.votesNeedSettlement`)),
+                ...insertIf(action === "vote" && votesResult === "no-death", this.$t(`${consequencesText}.votesNoDeath`)),
+                ...insertIf(action === "vote" && doesJudgeRequestAnotherVote, this.$t(`${consequencesText}.stutteringJudgeRequestedVote`)),
+                ...insertIf(action === "eat" && source.name === "werewolves" && targets[0].isInfected,
+                    this.$t(`${consequencesText}.piedPiperIsInfected`)),
+                ...insertIf(action === "mark" && targets.length,
+                    this.$t(`${consequencesText}.ravenMarked`, { markPenalty: options.roles.raven.markPenalty })),
+                ...insertIf(action === "choose-card" && areAllAdditionalCardsWerewolves(thiefAdditionalCards),
+                    this.$t(`${consequencesText}.thiefMustChooseCard`)),
+            ];
         },
     },
     methods: {